backend/requirements.txt
streamlit
python-dotenv
openai
google-generativeai
